optimize entity save in subgraph, fix rpc url, move env variable to config
