+++ app/.env.example
@@ -3,6 +3,7 @@ VITE_OTC_ADDRESS=0x9b4de7f9dBfDdc9E103dd5924300DaF7aAEF3cb6
 VITE_RPC=
 VITE_WS=
 VITE_THEGRAPH=https://gateway.thegraph.com/api/<API_KEY>/subgraphs/id/<SUBGRAPH_ID>
+VITE_HARMONY_RPC=https://api.harmony.one
 # or VITE_THEGRAPH=http://127.0.0.1:8000/subgraphs/name/otc-country for local test
 VITE_TLD=country
 VITE_TEST=1
