+++ app/src/graphql/client.ts
@@ -1,5 +1,6 @@
 import { GraphQLClient } from 'graphql-request'
+import * as config from '~/helpers/config'
 
-const client = new GraphQLClient(import.meta.env.VITE_THEGRAPH)
+const client = new GraphQLClient(config.THEGRAPH)
 
 export default client
