+++ app/src/helpers/link.ts
@@ -1,0 +1,7 @@
+import * as config from '~/helpers/config'
+
+export const host = config.TEST
+  ? 'sepolia.etherscan.io'
+  : 'explorer.harmony.one'
+
+export const hashLink = (hash: string): string => `https://${host}/tx/${hash}`
